fix: fix calendly test code
